requirements.txt
psycopg2-binary~=2.9.9
Flask~=2.2.5
Flask-Login~=0.6.2
Jinja2~=3.1.2
Werkzeug~=2.3.7
openai~=1.51.2
WTForms~=3.0.1
Flask-Migrate~=4.0.4
requests~=2.32.3
pillow~=10.4.0
pytesseract~=0.3.13
imutils~=0.5.4
opencv-python~=4.10.0.84